Baleen <-> JsonSchema supports IntegerType and NumericType Json should use BigDecimal and BigInteger for numbers, so using IntegerType and NumericType is better than using LongType and DoubleType

+++ jsonschema-model/src/main/kotlin/com/shoprunner/baleen/jsonschema/v4/IntegerSchema.kt
@@ -1,8 +1,10 @@
 package com.shoprunner.baleen.jsonschema.v4
 
+import java.math.BigInteger
+
 data class IntegerSchema(
-    val maximum: Long? = null,
-    val minimum: Long? = null
+    val maximum: BigInteger? = null,
+    val minimum: BigInteger? = null
 ) : JsonSchema() {
     val type = JsonType.integer
 }
