+++ jsonschema-model/src/main/kotlin/com/shoprunner/baleen/jsonschema/v4/NumberSchema.kt
@@ -1,8 +1,10 @@
 package com.shoprunner.baleen.jsonschema.v4
 
+import java.math.BigDecimal
+
 data class NumberSchema(
-    val maximum: Double? = null,
-    val minimum: Double? = null
+    val maximum: BigDecimal? = null,
+    val minimum: BigDecimal? = null
 ) : JsonSchema() {
     val type = JsonType.number
 }
